% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_common.R
\name{bord_dist}
\alias{bord_dist}
\title{Distance to SVM decision boundary}
\usage{
bord_dist(x, y, krn = "linear")
}
\arguments{
\item{x}{a data frame of input attributes.}

\item{y}{a factor vector with the output class of each sample.}

\item{krn}{a character with the kernel of SVM -see \code{e1071::svm} (default: "\code{linear}").}
}
\value{
A vector of length \code{nrow(x)} with the distance of each sample to the decision boundary.
}
\description{
Calculation of the distance of each sample to the SVM decision boundary in a classification problem.
}
\keyword{internal}
