% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_common.R
\name{bord_noise}
\alias{bord_noise}
\title{Mislabeling based on k-nearest neighbors}
\usage{
bord_noise(x, y, num_noise, idx_noise, k)
}
\arguments{
\item{x}{a data frame of input attributes.}

\item{y}{a factor vector with the output class of each sample.}

\item{num_noise}{an integer with the number of noisy samples.}

\item{idx_noise}{an integer vector with the indices of noisy samples.}

\item{k}{an integer with the number of nearest neighbors to use.}
}
\value{
A vector of length \code{length(y)} with the class of each sample, including the new noisy 
classes for the samples with indices \code{idx_noise}.
}
\description{
Computation of a noisy label based on majority class among \emph{k} nearest neighbors with different label.
}
\keyword{internal}
