% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/021_mulc_udir_ln.R
\name{mulc_udir_ln}
\alias{mulc_udir_ln}
\alias{mulc_udir_ln.default}
\alias{mulc_udir_ln.formula}
\title{Multiple-class unidirectional label noise}
\usage{
\method{mulc_udir_ln}{default}(x, y, level, goal, order = levels(y), sortid = TRUE, ...)

\method{mulc_udir_ln}{formula}(formula, data, ...)
}
\arguments{
\item{x}{a data frame of input attributes.}

\item{y}{a factor vector with the output class of each sample.}

\item{level}{a double in [0,1] with the noise level to be introduced.}

\item{goal}{an integer vector with the indices of noisy classes for each class.}

\item{order}{a character vector indicating the order of the classes (default: \code{levels(y)}).}

\item{sortid}{a logical indicating if the indices must be sorted at the output (default: \code{TRUE}).}

\item{...}{other options to pass to the function.}

\item{formula}{a formula with the output class and, at least, one input attribute.}

\item{data}{a data frame in which to interpret the variables in the formula.}
}
\value{
An object of class \code{ndmodel} with elements:
\item{xnoise}{a data frame with the noisy input attributes.}
\item{ynoise}{a factor vector with the noisy output class.}
\item{numnoise}{an integer vector with the amount of noisy samples per class.}
\item{idnoise}{an integer vector list with the indices of noisy samples.}
\item{numclean}{an integer vector with the amount of clean samples per class.}
\item{idclean}{an integer vector list with the indices of clean samples.}
\item{distr}{an integer vector with the samples per class in the original data.}
\item{model}{the full name of the noise introduction model used.}
\item{param}{a list of the argument values.}
\item{call}{the function call.}
}
\description{
Introduction of \emph{Multiple-class unidirectional label noise} into a classification dataset.
}
\details{
\emph{Multiple-class unidirectional label noise} introduction model randomly selects (\code{level}·100)\% of the samples
of each class \emph{c} with \code{goal}[c] != \code{NA}. Then, the labels \emph{c} of these samples are replaced by the class indicated in 
\code{goal}[c]. The order of indices in \code{goal} is determined by
\code{order}.
}
\note{
Noise model adapted from the papers in References.
}
\examples{
# load the dataset
data(iris2D)

# usage of the default method
set.seed(9)
outdef <- mulc_udir_ln(x = iris2D[,-ncol(iris2D)], y = iris2D[,ncol(iris2D)], level = 0.1, 
                        goal = c(NA, 1, 2), order = c("virginica", "setosa", "versicolor"))

# show results
summary(outdef, showid = TRUE)
plot(outdef)

# usage of the method for class formula
set.seed(9)
outfrm <- mulc_udir_ln(formula = Species ~ ., data = iris2D, level = 0.1, 
                        goal = c(NA, 1, 2), order = c("virginica", "setosa", "versicolor"))

# check the match of noisy indices
identical(outdef$idnoise, outfrm$idnoise)

}
\references{
Q. Wang, B. Han, T. Liu, G. Niu, J. Yang, and C. Gong. 
\strong{Tackling instance-dependent label noise via a universal probabilistic model}. 
In \emph{Proc. 35th AAAI Conference on Artificial Intelligence}, pages 10183-10191, 2021.
url:\url{https://ojs.aaai.org/index.php/AAAI/article/view/17221}.
}
\seealso{
\code{\link{minp_uni_ln}}, \code{\link{print.ndmodel}}, \code{\link{summary.ndmodel}}, \code{\link{plot.ndmodel}}
}
