% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_common.R
\name{noisetype}
\alias{noisetype}
\title{Type of noise introduced by a noise model}
\usage{
noisetype(model)
}
\arguments{
\item{model}{a character with the function name of the noise model.}
}
\value{
A character with the type of noise \code{model} introduces. It can be \code{cla} for 
label noise, \code{att} for attribute noise or \code{com} for combined noise.
}
\description{
Given the function name of a model, it returns the type of noise it 
introduces: label, attributes, or both.
}
\keyword{internal}
