% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndmodel.R
\name{print.sum.ndmodel}
\alias{print.sum.ndmodel}
\title{Print function for class sum.ndmodel}
\usage{
\method{print}{sum.ndmodel}(x, ...)
}
\arguments{
\item{x}{an object of class \code{sum.ndmodel}.}

\item{...}{other options to pass to the function.}
}
\value{
This function does not return any value.
}
\description{
Auxiliary function for printing information about the noise
introduction process contained in an object of class \code{sum.ndmodel}.
}
\keyword{internal}
