% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_common.R
\name{runif_replace}
\alias{runif_replace}
\title{Random numbers considering reference values}
\usage{
runif_replace(n, min, max, original, ref)
}
\arguments{
\item{n}{an integer with the amount of random numbers to generate.}

\item{min}{a double with the lower limit of the distribution.}

\item{max}{a double with the upper limit of the distribution.}

\item{original}{a boolean indicating if the values in \code{ref} can be chosen.}

\item{ref}{a double vector with \code{n} reference values.}
}
\value{
A double vector with the numbers generated.
}
\description{
Generate \code{n} random numbers following a uniform distribution 
between \code{min} and \code{max}. The values in \code{ref} can be chosen or not, 
according to \code{original}.
}
\keyword{internal}
