% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_common.R
\name{safe_sample}
\alias{safe_sample}
\title{Safe sample function}
\usage{
safe_sample(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{a vector with the alternatives to choose.}

\item{size}{an integer with the number of elements to select from \code{x}.}

\item{replace}{a boolean indicating if the elements should be chosen with replacement (default: \code{FALSE}).}

\item{prob}{a double vector with the probability associated to each element (default: \code{NULL}).}
}
\value{
A vector with the elements chosen.
}
\description{
Similar to standard \code{sample} function. Safe sample function considering the special case of an integer vector with only one element.
}
\keyword{internal}
