% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_common.R
\name{sample_replace}
\alias{sample_replace}
\title{Sample considering reference values}
\usage{
sample_replace(x, size, original, ref)
}
\arguments{
\item{x}{a vector with the alternatives to choose.}

\item{size}{an integer with the number of elements to select from \code{x}.}

\item{original}{a boolean indicating if the values in \code{ref} can be chosen.}

\item{ref}{a vector with \code{n} reference values.}
}
\value{
A vector with the elements chosen from \code{x}.
}
\description{
Similar to standard \code{sample} function. The values in \code{ref} can be chosen or not, 
according to \code{original}.
}
\keyword{internal}
