% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/036_sym_cuni_ln.R
\name{sym_cuni_ln}
\alias{sym_cuni_ln}
\alias{sym_cuni_ln.default}
\alias{sym_cuni_ln.formula}
\title{Symmetric completely-uniform label noise}
\usage{
\method{sym_cuni_ln}{default}(x, y, level, sortid = TRUE, ...)

\method{sym_cuni_ln}{formula}(formula, data, ...)
}
\arguments{
\item{x}{a data frame of input attributes.}

\item{y}{a factor vector with the output class of each sample.}

\item{level}{a double in [0,1] with the noise level to be introduced.}

\item{sortid}{a logical indicating if the indices must be sorted at the output (default: \code{TRUE}).}

\item{...}{other options to pass to the function.}

\item{formula}{a formula with the output class and, at least, one input attribute.}

\item{data}{a data frame in which to interpret the variables in the formula.}
}
\value{
An object of class \code{ndmodel} with elements:
\item{xnoise}{a data frame with the noisy input attributes.}
\item{ynoise}{a factor vector with the noisy output class.}
\item{numnoise}{an integer vector with the amount of noisy samples per class.}
\item{idnoise}{an integer vector list with the indices of noisy samples.}
\item{numclean}{an integer vector with the amount of clean samples per class.}
\item{idclean}{an integer vector list with the indices of clean samples.}
\item{distr}{an integer vector with the samples per class in the original data.}
\item{model}{the full name of the noise introduction model used.}
\item{param}{a list of the argument values.}
\item{call}{the function call.}
}
\description{
Introduction of \emph{Symmetric completely-uniform label noise} into a classification dataset.
}
\details{
\emph{Symmetric completely-uniform label noise} randomly selects (\code{level}·100)\% of the samples
in the dataset with independence of their class. Then, the labels of these samples are randomly
replaced by others within the set of class labels. Note that this model can choose the 
original label of a sample as noisy.
}
\note{
Noise model adapted from the papers in References.
}
\examples{
# load the dataset
data(iris2D)

# usage of the default method
set.seed(9)
outdef <- sym_cuni_ln(x = iris2D[,-ncol(iris2D)], y = iris2D[,ncol(iris2D)], level = 0.1)

# show results
summary(outdef, showid = TRUE)
plot(outdef)

# usage of the method for class formula
set.seed(9)
outfrm <- sym_cuni_ln(formula = Species ~ ., data = iris2D, level = 0.1)

# check the match of noisy indices
identical(outdef$idnoise, outfrm$idnoise)

}
\references{
A. Ghosh and A. S. Lan. \strong{Contrastive learning improves model robustness under label noise}. 
In \emph{Proc. 2021 IEEE Conference on Computer Vision and Pattern Recognition Workshops}, 
pages 2703-2708, 2021.
\doi{10.1109/CVPRW53098.2021.00304}.
}
\seealso{
\code{\link{sym_uni_ln}}, \code{\link{sym_cuni_an}}, \code{\link{print.ndmodel}}, \code{\link{summary.ndmodel}}, \code{\link{plot.ndmodel}}
}
