% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/049_sym_opt_ln.R
\name{sym_opt_ln}
\alias{sym_opt_ln}
\alias{sym_opt_ln.default}
\alias{sym_opt_ln.formula}
\title{Symmetric optimistic label noise}
\usage{
\method{sym_opt_ln}{default}(x, y, level, levelH = 0.9, order = levels(y), sortid = TRUE, ...)

\method{sym_opt_ln}{formula}(formula, data, ...)
}
\arguments{
\item{x}{a data frame of input attributes.}

\item{y}{a factor vector with the output class of each sample.}

\item{level}{a double in [0,1] with the noise level to be introduced.}

\item{levelH}{a double in (0.5, 1] with the noise level for higher classes (default: 0.9).}

\item{order}{a character vector indicating the order of the classes (default: \code{levels(y)}).}

\item{sortid}{a logical indicating if the indices must be sorted at the output (default: \code{TRUE}).}

\item{...}{other options to pass to the function.}

\item{formula}{a formula with the output class and, at least, one input attribute.}

\item{data}{a data frame in which to interpret the variables in the formula.}
}
\value{
An object of class \code{ndmodel} with elements:
\item{xnoise}{a data frame with the noisy input attributes.}
\item{ynoise}{a factor vector with the noisy output class.}
\item{numnoise}{an integer vector with the amount of noisy samples per class.}
\item{idnoise}{an integer vector list with the indices of noisy samples.}
\item{numclean}{an integer vector with the amount of clean samples per class.}
\item{idclean}{an integer vector list with the indices of clean samples.}
\item{distr}{an integer vector with the samples per class in the original data.}
\item{model}{the full name of the noise introduction model used.}
\item{param}{a list of the argument values.}
\item{call}{the function call.}
}
\description{
Introduction of \emph{Symmetric optimistic label noise} into a classification dataset.
}
\details{
\emph{Symmetric optimistic label noise} randomly selects (\code{level}·100)\% of the samples
in the dataset with independence of their class. 
In the optimistic case, the probability of a class \emph{i} of being mislabeled as class \emph{j} is 
higher for \emph{j} > \emph{i} in comparison to \emph{j} < \emph{i}.
Thus, when noise for a certain class occurs, it is assigned to a random higher class with probability \code{levelH} 
and to a random lower class with probability 1-\code{levelH}. The order of the classes is determined by 
\code{order}.
}
\note{
Noise model adapted from the papers in References.
}
\examples{
# load the dataset
data(iris2D)

# usage of the default method
set.seed(9)
outdef <- sym_opt_ln(x = iris2D[,-ncol(iris2D)], y = iris2D[,ncol(iris2D)], 
                     level = 0.1, order = c("virginica", "setosa", "versicolor"))

# show results
summary(outdef, showid = TRUE)
plot(outdef)

# usage of the method for class formula
set.seed(9)
outfrm <- sym_opt_ln(formula = Species ~ ., data = iris2D, 
                     level = 0.1, order = c("virginica", "setosa", "versicolor"))

# check the match of noisy indices
identical(outdef$idnoise, outfrm$idnoise)

}
\references{
R. C. Prati, J. Luengo, and F. Herrera. 
\strong{Emerging topics and challenges of learning from noisy data in nonstandard classification: 
a survey beyond binary class noise}. \emph{Knowledge and Information Systems}, 60(1):63–97, 2019.
\doi{10.1007/s10115-018-1244-4}.
}
\seealso{
\code{\link{sym_usim_ln}}, \code{\link{sym_natd_ln}}, \code{\link{print.ndmodel}}, \code{\link{summary.ndmodel}}, \code{\link{plot.ndmodel}}
}
