% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/072_uncs_guni_cn.R
\name{uncs_guni_cn}
\alias{uncs_guni_cn}
\alias{uncs_guni_cn.default}
\alias{uncs_guni_cn.formula}
\title{Unconditional/symmetric Gaussian/uniform combined noise}
\usage{
\method{uncs_guni_cn}{default}(x, y, level, k = 0.2, sortid = TRUE, ...)

\method{uncs_guni_cn}{formula}(formula, data, ...)
}
\arguments{
\item{x}{a data frame of input attributes.}

\item{y}{a factor vector with the output class of each sample.}

\item{level}{a double in [0,1] with the noise level to be introduced.}

\item{k}{a double in [0,1] with the scale used for the standard deviation (default: 0.2).}

\item{sortid}{a logical indicating if the indices must be sorted at the output (default: \code{TRUE}).}

\item{...}{other options to pass to the function.}

\item{formula}{a formula with the output class and, at least, one input attribute.}

\item{data}{a data frame in which to interpret the variables in the formula.}
}
\value{
An object of class \code{ndmodel} with elements:
\item{xnoise}{a data frame with the noisy input attributes.}
\item{ynoise}{a factor vector with the noisy output class.}
\item{numnoise}{an integer vector with the amount of noisy samples per variable.}
\item{idnoise}{an integer vector list with the indices of noisy samples per variable.}
\item{numclean}{an integer vector with the amount of clean samples per variable.}
\item{idclean}{an integer vector list with the indices of clean samples per variable.}
\item{distr}{an integer vector with the samples per class in the original data.}
\item{model}{the full name of the noise introduction model used.}
\item{param}{a list of the argument values.}
\item{call}{the function call.}
}
\description{
Introduction of \emph{Unconditional/symmetric Gaussian/uniform combined noise} into a classification dataset.
}
\details{
\emph{Unconditional/symmetric Gaussian/uniform combined noise} corrupts all the samples for
each attribute in the dataset. Their values are corrupted by adding a random value
following a Gaussian distribution of \emph{mean} = 0 and \emph{standard deviation} = (\emph{max}-\emph{min})·\code{k}, being
\emph{max} and \emph{min} the limits of the attribute domain. For nominal attributes, a random value is chosen. 
Additionally, this noise model also selects (\code{level}·100)\% of the samples
in the dataset with independence of their class. The labels of these samples are randomly
replaced by different ones within the set of class labels.
}
\note{
Noise model adapted from the papers in References.
}
\examples{
# load the dataset
data(iris2D)

# usage of the default method
set.seed(9)
outdef <- uncs_guni_cn(x = iris2D[,-ncol(iris2D)], y = iris2D[,ncol(iris2D)], level = 0.1)

# show results
summary(outdef, showid = TRUE)
plot(outdef)

# usage of the method for class formula
set.seed(9)
outfrm <- uncs_guni_cn(formula = Species ~ ., data = iris2D, level = 0.1)

# check the match of noisy indices
identical(outdef$idnoise, outfrm$idnoise)

}
\references{
S. Kazmierczak and J. Mandziuk. 
\strong{A committee of convolutional neural networks for image classification in the 
concurrent presence of feature and label noise}. 
In \emph{Proc. 16th International Conference on Parallel Problem Solving from Nature}, 
volume 12269 of LNCS, pages 498-511, 2020.
\doi{10.1007/978-3-030-58112-1_34}.
}
\seealso{
\code{\link{sym_cuni_cn}}, \code{\link{sym_cuni_an}}, \code{\link{print.ndmodel}}, \code{\link{summary.ndmodel}}, \code{\link{plot.ndmodel}}
}
